Updated the entity controller to check for permissions. Updated Entity Query Controller to include the default dacl if entity has no dacl.
Summary: Entity Controller should be able to check permissions and when querying the entities, it should include the default dacl if no dacl saved for that entity.

Test Plan: Run netric client and open entity and check for permissions.

Reviewers: sky

Differential Revision: https://src.aereusdev.com/D423

+++ src/Netric/Controller/EntityController.php
@@ -1,6 +1,7 @@
 <?php
 namespace Netric\Controller;
 
+use Netric\Entity\Entity;
 use Netric\EntityDefinition\Field;
 use Netric\Entity\EntityInterface;
 use Netric\Mvc;
@@ -9,6 +10,7 @@ use Netric\EntityDefinition\EntityDefinitionLoaderFactory;
 use Netric\Entity\FormsFactory;
 use Netric\Entity\BrowserView\BrowserViewServiceFactory;
 use Netric\Entity\EntityLoaderFactory;
+use Netric\Permissions\DaclLoaderFactory;
 use Netric\EntityDefinition\DataMapper\DataMapperFactory as EntityDefinitionDataMapperFactory;
 
 /**
@@ -159,9 +161,18 @@ class EntityController extends Mvc\AbstractAccountController
         if (isset($params['id']) && isset($params['obj_type'])) {
             // Retrieve the entity by id
             $entity = $loader->get($params['obj_type'], $params['id']);
-        }
 
-        // TODO: Check permissions
+            // If user is not allowed, then return an error
+            /*if (!$this->checkIfUserIsAllowed($dacl::PERM_VIEW)) {
+                return $this->sendOutput(
+                    array(
+                        "error" => "You dont have permission to view this.",
+                        "id" => $entity->getId(),
+                        "params" => $params
+                    )
+                );
+            }*/
+        }
 
         return $this->sendOutput(($entity) ? $entity->toArray() : []);
     }
@@ -256,8 +267,14 @@ class EntityController extends Mvc\AbstractAccountController
 
         foreach ($ids as $did) {
             $entity = $loader->get($objType, $did);
-            if ($dataMapper->delete($entity)) {
-                $ret[] = $did;
+
+            // Check first if we have permission to delete this entity
+            if ($this->checkIfUserIsAllowed($entity, $dacl::PERM_DELETE)) {
+
+                // Proceed with the deleting this entity
+                if ($dataMapper->delete($entity)) {
+                    $ret[] = $did;
+                }
             }
         }
 
@@ -814,5 +831,21 @@ class EntityController extends Mvc\AbstractAccountController
 
         // Return the groupings object
         return $groupings;
+    }
+
+    private function checkIfUserIsAllowed(Entity $entity, $permission) {
+
+        // Check entity permission
+        $daclLoader = $this->account->getServiceManager()->get(DaclLoaderFactory::class);
+        $dacl = $daclLoader->getForEntity($entity);
+        $user = $this->account->getUser();
+        $isOwner = ($entity->getValue("creator_id") == $user->getId());
+
+        // If the current user wants to retrieve its own entity, then we will allow it
+        if ($this->getObjType() === "user" && $entity->getId() == $user->getId()) {
+            $isOwner = true;
+        }
+
+        return $dacl->isAllowed($user, $permission, $isOwner);
     }
 }
